Revert ListPart workaround, rename "build_display" to "shape_build_display".
Update the doc, concatenate 2 sections of "Shape Tags" into only one.
